+++ src/crudify/interface/controllermethods.type.ts
@@ -4,6 +4,7 @@ export type ControllerMethods =
   | "findAll"
   | "findOne"
   | "put"
+  | "putBulk"
   | "update"
   | "updateBulk"
   | "delete"